fix(types): resolve 321 mypy errors with targeted type fixes across 67 files
- Add explicit casts (int/str/float/bool/dict) for no-any-return errors
- Narrow logger types from Logger|StructuredLogger to StructuredLogger
- Fix var-annotated errors with proper type annotations
- Add type: ignore[import-untyped] for untyped library imports (yaml, chess, etc.)
- Cast LLMResponse in base agent to resolve union return type issues
- Fix assignment errors with proper Optional typing and narrowing
- Add return type annotations to override methods matching parent signatures

Mypy errors reduced from 367 to 46 (87% reduction)

+++ src/neuro_symbolic/state.py
@@ -338,7 +338,7 @@ class NeuroSymbolicState:
 
         # Combine
         neural_weight = 1.0 - symbolic_weight
-        return symbolic_weight * symbolic_sim + neural_weight * neural_sim
+        return float(symbolic_weight * symbolic_sim + neural_weight * neural_sim)
 
     def to_dict(self) -> dict[str, Any]:
         """Convert state to dictionary for serialization."""
